Rework applying speedhack
The speedhack is now applied during "change", but at the start based on the key rather than heuristically finding a likely memory address.

The speedhack applies a pitch shift effect so songs are now at normal pitch.

After changing the speed percent you now have to go into practice mode so that change is called with the correct keys.  This needs to be fixed still.